Move UnnecessaryChainedSafeCall to style
