+++ detekt-rules-standard-library/src/main/kotlin/dev/detekt/rules/standardlibrary/StandardLibraryProvider.kt
@@ -35,7 +35,6 @@ class StandardLibraryProvider : DefaultRuleSetProvider {
                 ::TrimMultilineRawString,
                 ::UnnecessaryAny,
                 ::UnnecessaryApply,
-                ::UnnecessaryChainedSafeCall,
                 ::UnnecessaryFilter,
                 ::UnnecessaryLet,
                 ::UnnecessaryReversed,
