+++ detekt-rules-style/src/main/kotlin/dev/detekt/rules/style/StyleGuideProvider.kt
@@ -95,6 +95,7 @@ class StyleGuideProvider : DefaultRuleSetProvider {
                 ::MultilineLambdaItParameter,
                 ::StringShouldBeRawString,
                 ::UnnecessaryBackticks,
+                ::UnnecessaryChainedSafeCall,
                 ::MaxChainedCallsOnSameLine,
                 ::DoubleNegativeLambda,
                 ::DoubleNegativeExpression,
